% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_or.R
\name{add_or}
\alias{add_or}
\title{Function for transform a largecollapsedVcf into a dataframe with
predicted states (only with chr,start,end and metadatacolumn)}
\usage{
add_or(start_coord, end_coord, seqnames, group, split_vcf_df = NULL)
}
\arguments{
\item{start_coord}{start chromosomal position of every block}

\item{end_coord}{end chromosomal position of every block}

\item{seqnames}{name of the chromosome of the blocks (e.g. chr1 or 1)}

\item{group}{hidden state assigned by  UPDHmm}

\item{split_vcf_df}{input vcf file}
}
\value{
dataframe
}
\description{
Function for transform a largecollapsedVcf into a dataframe with
predicted states (only with chr,start,end and metadatacolumn)
}
