% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_viterbi.R
\name{apply_viterbi}
\alias{apply_viterbi}
\title{Apply the hidden Markov model using the Viterbi algorithm.}
\usage{
apply_viterbi(largecollapsedVcf = NULL, hmm = NULL, genotypes = NULL)
}
\arguments{
\item{largecollapsedVcf}{input vcf file}

\item{hmm}{Hidden Markov Model used to infer the events}

\item{genotypes}{Possible GT formats and its correspondency with the hmm}
}
\value{
largecollapsedVcf
}
\description{
Apply the hidden Markov model using the Viterbi algorithm.
}
