% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df_vcf.R
\name{as_df_vcf}
\alias{as_df_vcf}
\title{Function for transform a largecollapsedVcf into a dataframe with predicted
states (only with chr,start,end and metadatacolumn)}
\usage{
as_df_vcf(largecollapsedVcf)
}
\arguments{
\item{largecollapsedVcf}{Vcf file}
}
\value{
dataframe
}
\description{
Function for transform a largecollapsedVcf into a dataframe with predicted
states (only with chr,start,end and metadatacolumn)
}
