% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_vcf.R
\name{blocks_vcf}
\alias{blocks_vcf}
\title{Function for simplify into blocks contiguous variants with same state}
\usage{
blocks_vcf(df)
}
\arguments{
\item{df}{df result from as_df_vcf function}
}
\value{
dataframe
}
\description{
Function for simplify into blocks contiguous variants with same state
}
