% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmm}
\alias{hmm}
\title{HMM data with alelle frequencies in different states}
\format{
A list with 5 different elements
\describe{
\item{states}{5 different possible states}
\item{symbols}{code symbol use for the genotype's combination}
\item{startProbs}{the initial probabilites of every state}
\item{transProbs}{probability between changing from one state to another}
\item{emsissionProbs}{given a certain genotype's combination what are the odds of every possible state}
}
}
\source{
Created in house for calculating Uniparental disomy events
}
\usage{
data(hmm)
}
\description{
It contains the emission probabilites, transmission probabilites ...
}
\examples{
data(hmm)
}
\keyword{datasets}
