% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPDhmm-package.R
\docType{package}
\name{UPDhmm-package}
\alias{UPDhmm}
\alias{UPDhmm-package}
\title{UPDhmm: Detecting Uniparental Disomy through NGS trio data}
\description{
<<<<<<< HEAD
Uniparental disomy (UPD) is a genetic condition where an individual inherits both copies of a chromosome or part of it from one parent, rather than one copy from each parent. This package contains a HMM for detecting UPDs through HTS (High Throughput Sequencing) data from trio assays. By analyzing the genotypes in the trio, the model infers a hidden state (normal, father isodisomy, mother isodisomy, father heterodisomy and mother heterodisomy).
=======
Uniparental disomy (UPD) is a genetic condition where an individual inherits both copies of a chromosome or part of it from one parent, rather than one copy from each parent. This package contains a HMM for detecting UPDs throguh HTS (High Throughput Sequencing) data from trio assays. By analizing the genotypes in the trio, the model infers a hidden state (normal, father isodisomy, mother isodisomy, father heterodisomy and mother heterodisomy).
>>>>>>> upstream/devel
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/martasevilla/UPDhmm}
  \item Report bugs at \url{https://github.com/martasevilla/UPDhmm/issues}
}

}
\author{
\strong{Maintainer}: Marta Sevilla \email{marta.sevilla@upf.edu} (\href{https://orcid.org/0009-0005-0179-920X}{ORCID})

Authors:
\itemize{
  \item Carlos Ruiz-Arenas \email{cruizarenas@unav.es} (\href{https://orcid.org/0000-0002-6014-3498}{ORCID})
}

}
\keyword{internal}
