% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_check.R
\name{vcf_check}
\alias{vcf_check}
\title{check quality parameters and change ids}
\usage{
vcf_check(
  path_vcf,
  check_quality = NULL,
  father = NULL,
  mother = NULL,
  proband = NULL
)
}
\arguments{
\item{path_vcf}{The file in largeCollapsed format}

\item{check_quality}{If desired, quality parameters can be measured}

\item{father}{Introduce father's name sample}

\item{mother}{Introduce mother's name sample}

\item{proband}{Introduce proband's name sample}
}
\value{
LargecollapdsedVCf (VariantAnnotation vcf format)
}
\description{
This function will take the vcf (must be gz with an index file and will
convert it into a large collapsedVcf object using variantanntoation package)
it will have one additional parameter, quality_check, that will give warnigns
if there are positions with no good quality.
}
\examples{
fl <- system.file("extdata", "test.vcf.gz", package = "UPDhmm")
vcf <- VariantAnnotation::readVcf(fl)
vcf_check(vcf, proband = "Sample1", mother = "Sample3", father = "Sample2")






}
