% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_or.R
\name{add_or}
\alias{add_or}
\title{Function to transform a large collapsed VCF into a dataframe,
incorporating predicted states along with the log-likelihood
ratio and p-value.}
\usage{
add_or(
  filtered_def_blocks_states = NULL,
  largecollapsedVcf = NULL,
  hmm = NULL,
  genotypes = NULL
)
}
\arguments{
\item{filtered_def_blocks_states}{data.frame object containing the blocks}

\item{largecollapsedVcf}{Input VCF file}

\item{hmm}{Hidden Markov Model used to infer the events}

\item{genotypes}{Possible GT formats and its correspondency with the hmm}
}
\value{
data.frame containing the transformed information.
}
\description{
Function to transform a large collapsed VCF into a dataframe,
incorporating predicted states along with the log-likelihood
ratio and p-value.
}
