% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_viterbi.R
\name{apply_viterbi}
\alias{apply_viterbi}
\title{This function is for applying the hidden-markov-model using the
viterbi algorithm
IMPORTANT: vector_samples should be previously created}
\usage{
apply_viterbi(largecollapsedVcf)
}
\arguments{
\item{largecollapsedVcf}{input vcf file}
}
\value{
largecollapsedVcf
}
\description{
This function is for applying the hidden-markov-model using the
viterbi algorithm
IMPORTANT: vector_samples should be previously created
}
