% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyViterbi.R
\name{applyViterbi}
\alias{applyViterbi}
\title{Apply the hidden Markov model using the Viterbi algorithm.}
\usage{
applyViterbi(largeCollapsedVcf, hmm)
}
\arguments{
\item{largeCollapsedVcf}{input vcf file}

\item{hmm}{Hidden Markov Model used to infer the events}
}
\value{
the input largeCollapsedVcf object updated with a new metadata column \emph{states},
which contains the inferred hidden state for each variant.
}
\description{
Apply the hidden Markov model using the Viterbi algorithm.
}
\keyword{internal}
