% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyViterbi.R
\name{applyViterbi}
\alias{applyViterbi}
\title{Apply the hidden Markov model using the Viterbi algorithm.}
\usage{
applyViterbi(largeCollapsedVcf, hmm)
}
\arguments{
\item{largeCollapsedVcf}{input vcf file}

\item{hmm}{Hidden Markov Model used to infer the events}
}
\value{
largeCollapsedVcf
}
\description{
This function takes a collapsed VCF object and a pre-defined Hidden Markov Model (HMM),
applies the Viterbi algorithm to infer the most likely sequence of hidden states for
each sample, and stores the resulting states as a new metadata column in the VCF object.
}
