% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyViterbi.R
\name{applyViterbi}
\alias{applyViterbi}
\title{Apply the hidden Markov model using the Viterbi algorithm.}
\usage{
applyViterbi(largeCollapsedVcf, hmm)
}
\arguments{
\item{largeCollapsedVcf}{input vcf file}

\item{hmm}{Hidden Markov Model used to infer the events}

\item{genotypes}{Possible GT formats and its correspondence with the hmm}
}
\value{
largeCollapsedVcf
}
\description{
Apply the hidden Markov model using the Viterbi algorithm.
}
