% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDfVcf.R
\name{asDfVcf}
\alias{asDfVcf}
\title{Function to transform a large collapsed VCF into a dataframe with
predicted states, including chromosome, start position, end position
and metadata.}
\usage{
asDfVcf(largecollapsedVcf = NULL, genotypes = NULL)
}
\arguments{
\item{largecollapsedVcf}{Name of the large collapsed VCF file.}

\item{genotypes}{Possible GT formats and its correspondency with the hmm}
}
\value{
dataframe
}
\description{
Function to transform a large collapsed VCF into a dataframe with
predicted states, including chromosome, start position, end position
and metadata.
}
