% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_events.R
\name{calculate_events}
\alias{calculate_events}
\title{Calculate UPD events in trio VCFs.}
\usage{
calculate_events(largecollapsedVcf = NULL, genotypes = NULL)
}
\arguments{
\item{largecollapsedVcf}{The VCF file in the general format}

\item{genotypes}{Possible GT formats and its correspondency with the hmm
(largecollapsedVcf) with VariantAnnotation package.}

\item{hmm}{Hidden Markov Model used to infer the events}
}
\value{
Dataframe object containing blocks of predicted events.
}
\description{
This function predicts the hidden states by applying the Viterbi algorithm
using the Hidden Markov Model (HMM) from the UPDhmm package. It takes the
genotypes of the trio as input and includes a final step to simplify the
results into blocks.
}
\examples{
fl <- system.file("extdata", "test.vcf.gz", package = "UPDhmm")
vcf <- VariantAnnotation::readVcf(fl)
largecollapsedVcf <- vcf_check(vcf,
  proband = "Sample1", mother = "Sample3",
  father = "Sample2"
)
calculate_events(largecollapsedVcf)
}
