% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_events.R
\name{calculate_events}
\alias{calculate_events}
\title{Calculate UPD events in trio vcfs}
\usage{
calculate_events(largecollapsedVcf)
}
\arguments{
\item{largecollapsedVcf}{The general format of vcf with VariantAnnotation
package}
}
\value{
optional save object (raw and simplify ) and export bed file
}
\description{
This function is for appliying the hidden-markov-model using the viterbi
algorithm after that it simplify into blocks and finally export into a
bed file
}
\examples{
fl <- system.file("extdata", "test.vcf.gz", package = "UPDhmm")
vcf <- VariantAnnotation::readVcf(fl)
largecollapsedVcf <- vcf_check(vcf,
  proband = "Sample1", mother = "Sample3",
  father = "Sample2"
)
calculate_events(largecollapsedVcf)
}
