% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEvents.R
\name{calculateEvents}
\alias{calculateEvents}
\title{Calculate UPD events in trio VCFs.}
\usage{
calculateEvents(largeCollapsedVcf, hmm = NULL)
}
\arguments{
\item{largeCollapsedVcf}{The VCF file in the general format
(largeCollapsedVcf) with VariantAnnotation package. Previously edited with
\code{vcfCheck()} function from UPDhmm package}

\item{hmm}{Default = NULL. If no arguments are added, the package
will use the default HMM already implemented, based on Mendelian
inheritance. If an optional HMM is desired, it should adhere to the
general HMM format from \code{HMM} package with the following elements inside
a list:
\enumerate{
\item The hidden state names in the "States" vector.
\item All possible observations in the "Symbols" vector.
\item Start probabilities of every hidden state in the "startProbs" vector.
\item Transition probabilities matrix between states in "transProbs".
\item Probabilities associated between every hidden state and all possible
observations in the "emissionProbs" matrix.
}}
}
\value{
A data.frame object containing all detected events in the provided trio.
If no events are found, the function  will return an empty data.frame.
}
\description{
This function predicts the hidden states by applying the Viterbi algorithm
using the Hidden Markov Model (HMM) from the UPDhmm package. It takes the
genotypes of the trio as input and includes a final step to simplify the
results into blocks.
}
\examples{
file <- system.file(package = "UPDhmm", "extdata", "test_het_mat.vcf.gz")
vcf <- VariantAnnotation::readVcf(file)
processedVcf <- vcfCheck(vcf,
    proband = "NA19675", mother = "NA19678",
    father = "NA19679"
)

}
