% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEvents.R
\name{calculateEvents}
\alias{calculateEvents}
\title{Calculate UPD events in trio VCFs.}
\usage{
<<<<<<< HEAD
calculateEvents(largeCollapsedVcf, hmm = "hmm", genotypes)
}
\arguments{
\item{largeCollapsedVcf}{The VCF file in the general format}

\item{hmm}{Hidden Markov Model used to infer the events. The format should
adhere to the general HMM format with a series of elements:
\enumerate{
\item The hidden states names in the "States" vector.
\item All possible observations in the "Symbols" vector.
\item Start probabilities of every hidden state in the "startProbs" vector.
\item Transition probabilities matrix of the hidden states in "transProbs".
\item Probabilities associated between every hidden state and all possible
observations in the "emissionProbs" matrix.
}}

\item{genotypes}{Possible GT formats and its correspondence with the hmm
(largeCollapsedVcf) with VariantAnnotation package.}
=======
calculateEvents(largecollapsedVcf = NULL, hmm = "hmm", genotypes = NULL)
}
\arguments{
\item{largecollapsedVcf}{The VCF file in the general format}

\item{hmm}{hmm constructed for identifying hidden states}

\item{genotypes}{Possible GT formats and its correspondency with the hmm
(largecollapsedVcf) with VariantAnnotation package.}
>>>>>>> upstream/devel
}
\value{
Dataframe object containing blocks of predicted events.
}
\description{
This function predicts the hidden states by applying the Viterbi algorithm
using the Hidden Markov Model (HMM) from the UPDhmm package. It takes the
genotypes of the trio as input and includes a final step to simplify the
results into blocks.
}
\examples{
file <- system.file(package = "UPDhmm", "extdata", "test_het_mat.vcf.gz")
vcf <- VariantAnnotation::readVcf(file)
<<<<<<< HEAD
processedVcf <- vcfCheck(vcf,
    proband = "NA19675", mother = "NA19678",
    father = "NA19679"
)
=======
processedVcf <- vcfCheck(vcf,proband = "NA19675",mother = "NA19678",
father = "NA19679")
>>>>>>> upstream/devel

calculateEvents(processedVcf)
}
