% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEvents.R
\name{calculateEvents}
\alias{calculateEvents}
\title{Calculate UPD events in trio VCFs.}
\usage{
calculateEvents(
  largeCollapsedVcf,
  hmm = NULL,
  genotypes = c(`0/0` = "1", `0/1` = "2", `1/0` = "2", `1/1` = "3", `0|0` = "1", `0|1` =
    "2", `1|0` = "2", `1|1` = "3")
)
}
\arguments{
\item{largeCollapsedVcf}{The VCF file in the general format}

\item{hmm}{Optional argument. Hidden Markov Model used to infer the events.
The format should
adhere to the general HMM format with a series of elements:
\enumerate{
\item The hidden states names in the "States" vector.
\item All possible observations in the "Symbols" vector.
\item Start probabilities of every hidden state in the "startProbs" vector.
\item Transition probabilities matrix of the hidden states in "transProbs".
\item Probabilities associated between every hidden state and all possible
observations in the "emissionProbs" matrix. #' If NULL, hmm applied would
be the default model
}}

\item{genotypes}{Optional argument. Possible GT formats and its
correspondence with the hmm
(largeCollapsedVcf) with VariantAnnotation package.}
}
\value{
Dataframe object containing blocks of predicted events.
}
\description{
This function predicts the hidden states by applying the Viterbi algorithm
using the Hidden Markov Model (HMM) from the UPDhmm package. It takes the
genotypes of the trio as input and includes a final step to simplify the
results into blocks.
}
\examples{
file <- system.file(package = "UPDhmm", "extdata", "test_het_mat.vcf.gz")
vcf <- VariantAnnotation::readVcf(file)
processedVcf <- vcfCheck(vcf,
    proband = "NA19675", mother = "NA19678",
    father = "NA19679"
)

calculateEvents(processedVcf)
}
