% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processChromosome.R
\name{processChromosome}
\alias{processChromosome}
\title{Process a single chromosome for UPD detection}
\usage{
processChromosome(
  vcf_chr,
  hmm,
  add_ratios = FALSE,
  field_DP = NULL,
  total_mean = NULL,
  mendelian_error_values
)
}
\arguments{
\item{vcf_chr}{CollapsedVCF object for one chromosome}

\item{hmm}{Hidden Markov Model object}

\item{genotypes}{Named vector mapping genotype strings to numeric states}
}
\value{
A data.frame of detected blocks for the chromosome, or NULL if error
}
\description{
Internal helper function to run the full pipeline on one chromosome:
\itemize{
\item applyViterbi
\item asDfVcf
\item blocksVcf
}
}
