% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfCheck.R
\name{vcfCheck}
\alias{vcfCheck}
\title{Check variant quality, rename trio samples, and encode genotypes numerically}
\usage{
vcfCheck(largeCollapsedVcf, father, mother, proband, check_quality = FALSE)
}
\arguments{
\item{largeCollapsedVcf}{A VCF object of class \code{CollapsedVCF}.}

\item{father}{Character string with the sample ID corresponding to the father.}

\item{mother}{Character string with the sample ID corresponding to the mother.}

\item{proband}{Character string with the sample ID corresponding to the proband.}

\item{check_quality}{Logical; if \code{TRUE}, basic quality warnings based on
read depth (DP) and genotype quality (GQ) are displayed.
Default: \code{FALSE}.}
}
\value{
A \code{CollapsedVCF} object identical to the input but with:
\itemize{
\item standardized sample names,
\item \code{colData(vcf)$ID} storing original sample names,
\item \code{mcols(vcf)$geno_coded} containing the trio genotype encoding.
}
}
\description{
This function processes a \code{CollapsedVCF} object by standardizing trio sample
names ("father", "mother", "proband"), optionally checking variant quality
metrics, and generating a numeric encoding of the trio genotypes for each
variant. The encoded genotypes are stored in the metadata column \code{geno_coded}.
}
\examples{
file <- system.file(package = "UPDhmm", "extdata", "test_het_mat.vcf.gz")
vcf  <- VariantAnnotation::readVcf(file)

processedVcf <- vcfCheck(
    vcf,
    proband = "NA19675",
    mother  = "NA19678",
    father  = "NA19679"
)

}
