% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfCheck.R
\name{vcfCheck}
\alias{vcfCheck}
\title{Check quality parameters (optional) and change IDs.}
\usage{
vcfCheck(
  largecollapsedVcf,
  check_quality = FALSE,
  father = NULL,
  mother = NULL,
  proband = NULL
)
}
\arguments{
\item{largecollapsedVcf}{The file in largeCollapsed format.}

\item{check_quality}{TRUE/FALSE. If quality parameters want to be measured.
Default = FALSE}

\item{father}{Name of the father's sample.}

\item{mother}{Name of the mother's sample.}

\item{proband}{Name of the proband's sample.}
}
\value{
LargecollapsedVCF (VariantAnnotation VCF format).
}
\description{
This function takes a VCF file and converts it into a large collapsedVcf
object using the VariantAnnotation package. It includes an optional
parameter, quality_check, which issues warnings if positions lack good
quality based on RD and GQ parameters in the input VCF.
}
\examples{
fl <- system.file("extdata", "test_het_mat.vcf.gz", package = "UPDhmm")
vcf <- VariantAnnotation::readVcf(fl)
processedVcf<-
vcfCheck(vcf, proband = "Sample1", mother = "Sample3", father = "Sample2")






}
