% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfCheck.R
\name{vcfCheck}
\alias{vcfCheck}
\title{Check variant quality (optional), rename samples, and numerically encode genotypes}
\usage{
vcfCheck(largeCollapsedVcf, father, mother, proband, check_quality = FALSE)
}
\arguments{
\item{largeCollapsedVcf}{The file in largeCollapsedVcf format.}

\item{father}{Name of the father's sample.}

\item{mother}{Name of the mother's sample.}

\item{proband}{Name of the proband's sample.}

\item{check_quality}{Optional argument. TRUE/FALSE. If quality parameters
want to be measured.
Default = FALSE}
}
\value{
largeCollapsedVcf (VariantAnnotation VCF format).
}
\description{
This function processes a VCF file by converting it into a \code{CollapsedVCF} object
using the VariantAnnotation package. It renames the samples to standard names
("father", "mother", "proband") for subsequent UPDhmm analysis, optionally
evaluates variant quality based on read depth (DP) and genotype quality (GQ),
and creates a numeric encoding of the trio genotypes in the metadata column \code{geno_coded}.
}
\details{
The numeric genotype encoding \code{geno_coded} is a 3-character string per variant,
representing the genotypes of father, mother, and proband respectively, where:
1 = homozygous reference, 2 = heterozygous, 3 = homozygous alternate.
}
\examples{
file <- system.file(package = "UPDhmm", "extdata", "test_het_mat.vcf.gz")
vcf <- VariantAnnotation::readVcf(file)
processedVcf <- vcfCheck(vcf,
    proband = "NA19675", 
    mother = "NA19678",
    father = "NA19679"
)

}
